refactor Project Structure
