+++ docs/swagger/component/response.js
