+++ docs/swagger/component/securityScheme.js
