+++ src/pkg/helpers/imgUrlGenerator.js
