+++ src/pkg/helpers/otpCodeGenerator.js
